Move the kernel launch to the analyze function
